+++ lib/generators/action_policy/policy/policy_generator.rb
@@ -7,7 +7,10 @@ module ActionPolicy
     class PolicyGenerator < ::Rails::Generators::NamedBase
       source_root File.expand_path("templates", __dir__)
 
-      invoke "action_policy:install"
+      def run_install_if_needed
+        return if ::Rails.root.join("app/policies/application_policy.rb").exist?
+        generate "action_policy:install"
+      end
 
       def create_policy
         template "policy.rb", File.join("app/policies", class_path, "#{file_name}_policy.rb")